🚧(chesspiece) Adding En Passant to move
I know what is wrong! The tests just try to do a regular move without regarding the additional field that I have added. My fix for this is that even though the test move and the en passant move are equal, I will check my move object for the en passant flag, since the test definitely doesn't have it.